backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.0
gunicorn==21.2.0
